Move CLI options to `opts.rs`
Only modifications to the options are making fields public where necessary for the main code.
Required to be able to use later for completions.
